Add new NullDataStore
I don't know how to fix the TOCTOU problem in the disk
based store at this time. Therefore to fix the production
explosions we'll use a datastore that doesn't actually
do anything because we wont be in danger of losing any
data if the listener fails to process a message.

Change-Id: I8f2c068805279dee11bf1b831f38121e888bb2c1
NOTE: Only because we don't currently use Adyen!