Configurable queue write mirroring
MultiQueueWriter is a disposable wrapper to make it possible to switch from
writing to ActiveMQ only, to both ActiveMQ and php-queue, to php-queue
only, just by changing configuration. The intent is that writes to ActiveMQ
change as little as possible.

Once activeMQ is decommissioned, we may want to get rid of this and push
arrays directly from consuming classes, assuming everything implements
Fifo

Bug: T141485
Change-Id: Ia23ef30d6851066461440851873b9570e4916bfe