DRY finding OrderReferenceId, get_class() -> instanceOf
Refactor AmazonApi and AmazonMessage for less duplicated code
(in the future).

There are a lot of places we might want the order reference ID,
which is the first 19 characters of a the derived IDs. Also, the
client setup has to happen for any API calls, so make in an
instance variable. Finally, use instanceOf instead of get_class
for less brittleness.

Bug: T147973
Change-Id: Ia074b1d9a475f3dfed5b25d09d77a3fc23197228