Fix Amazon order IDs for manual captures
When a capture that we request via the front end goes through
correctly, the associated IPN message has our merchant reference
as 'CaptureReferenceId'. However, when the capture fails and DS
has to push it through via the payments console, it gets an
Amazon-generated CaptureReferenceId starting with AUTHORIZE_.
That ID is useless to us, but we can look up the original
merchant reference via the AmazonCaptureId.

The order reference ID is the first 19 digits of that number,
and we should now be setting the merchant reference on the parent
order ID.

Bug: T147973
Change-Id: I02f54e8f74f07e436a3d25409c5377ed08c68b21