Fix insecure uncaught exception logging
When there's an uncaught error in a maintenance script, the last
chance error handler was smushing all the exception info into a
string and calling MaintenanceBase->error with $fatal set to true.

With $fatal == true, it logs the whole thing including stack trace
at alert level then exits. This patch fixes it to use the more
discerning exception logging that keeps the stack trace out of
failmails.

Bug: T155757
Change-Id: I23d28f722a0bb5675425e0a56a2366bed916c22e