Override configuration with specific instances
Homebrew dependency injection is nearly complete. Last thing we'd
need would be a way to specify a runnable to return instances for
nodes.

This is mostly just useful for test cases, where you want to create
a mock on the fly and configure some return values. Should let us
get rid of the silly mock types in the Amazon SDK fork.

Change-Id: I6445c01429a337482d52399c5dafb2fd32c50717