Cache bank lookups via PSR-6
Includes a simple cache implementation. Adds a new config key 'cache'
at the root of the default view, which should resolve to an object
that implements Psr\Cache\CacheItemPoolInterface.

With this, any project that depends on SmashPig can make its cache
available to SmashPig classes, as long as there is a PSR6 compatible
interface.

Bug: T128692
Change-Id: I457fb093c3777406efbfd4821acac47520af60d2