Work around Memcache not expiring some things
Memcache sometimes holds things too long when given unix timestamps for
expiry. This adds an application-level check for expiration in the one
place we're currently using the cache. If we need to use the cache in
other places before we figure out the Memcache issue, we should move
this logic to Core\Cache.

Bug: T128692
Change-Id: I0cd23d8dc7d97e5f7e7628215f8d35c7da250391