Hack to get availability straight from iDEAL
Hits an undocumented API that they use to show this page:
http://availability.ideal.nl/
Gets us bank names, codes, and percent availability. Adds a configuration
setting to drop banks from the list when they're under a certain
percent available.

Also reduces lookup frequency to every 15 minutes since we're not
hitting an official API endpoint.

TODO: If noone vetoes this, maybe put it in a different class.

Bug: T161153
Change-Id: I0d24f86a0ca2bd89337e5d3acec723db05cb69f0