Split provider configuration out of main
Each gets its own dir because they'll eventually have all the settings
from DonationInterface's config directories.

This is just a transitional commit, on the way to fully splitting out a
ProviderConfiguration class from GlobalConfiguration. Eventually we'll
be able to load multiple provider configurations at the same time, so
for example we can choose a provider based on capabilities.

Change-Id: I6009e97e9875feea584c7baaa129bb3aea97ddd5