Actually split out config classes
Context now holds a GlobalConfiguration object required for init, as well
as a ProviderConfiguration object that can be reset later. If none is provided
at startup, the context creates a default one.

ProviderConfiguration will eventually pull in all the auxiliary files
that DonationInterface reads from the gateway-specific config directories.

For tests, this means the base case can initialize the context in setup with a
single testing configuration object which points all queue and db objects at
in-memory datastructres.

TODO: Logger should react to setting provider configuration.

Bug: T143831
Change-Id: Ib5c810b0fd08e8c0147aaac1f75dda018f9723d2