Add .editorconfig
I couldn't actually find in the MediaWiki standard that it says we should use
tabs instead of spaces, but that is what is happening in this module.

Shame MediaWiki doesn't use psr-2 - would be one less, but like drupal they
probably chose a standard before psr-2 won the standards war.

I was amused to read the page as they advocated various things as 'being more
readable' - when people clearly have differing opinions on that.

Anyway, this isn't working in my IDE as well as I might hope with
the spaces vs tabs, but I'm pretty sure it is right & the magic is in
getting the IDE to work right (or waiting for a release that does).

Bug: T177725
Change-Id: I317c068b28dd2cdede9d854e187fb952ac0fe3e2