Send Amazon capture messages to jobs-amazon
Use the RecordPaymentJob introduced in the previous patch

TODO: Would be nice to have a test that runs the whole action chain,
then runs the RecordPaymentsJob. Would need to write an actual class
to replace the IPNHandler, since we're not using $config->object()
to get that and thus can't stuff the node with a mock object.

Bug: T194317
Change-Id: I7221d9da411f9ec06524d45babc71cb61023209a