Tag txn_type cart and payment_type instant as EC
It's not totally clear where these are coming from (they generally
have no payments in the front-end logs with matching date & amount),
but when the same txn ids show up in audit files later, they have
the Payment Source column filled out as "Express Checkout". Tag
them with the same gateway here so we don't file a double donation
in CiviCRM.

This seems to happen with about 1 in 10,000 PayPal transactions.

Bug: T230956
Change-Id: I97e625f218f3493ff006e33540442fa7af4c3357