Don't crash on new D*Local audit row types
This patch doesn't actually do anything with the data, it just
makes sure we don't blow up.

The Civi code which uses this parser doesn't have any way to update
refunds that have already been recorded, nor to records fees on
refunds, so we will need to update that eventually and then we can
revisit doing something here.

Bug: T236524
Change-Id: Ide0ce4117381420fec3087c1463555881561ba15