Fix Ingenico audit parsing for BPay donations
It's a really weird record, missing a bunch of the indicators that
we usually get. This patch adds a couple of workarounds, including
looking in another field for the date, pinning the gateway value to
'globalcollect' for legacy-only methods, and looking in another
field if the order_id is set to zero.

Bug: T245749
Change-Id: I087871dc1ac8accfe6fefeb8ed200cc62fe46355