Ingenico: don't throw ApiError when top-level has 'errors'
I initially thought the top-level 'errors' element was reserved for
problems with API authentication or other serious systematic issues.
It turns out that normal payment declines get a top-level 'errors'
key set as well, I guess for convenience. If we throw an ApiException
for those it's problematic and doesn't let us normalize the errors
for the consuming code.

Bug: T243340
Change-Id: Ib493c6556efb2a1768a6f2f6515d71f874b5dc14