Push most response properties to base, add requiresApproval
It felt wrong for cancel and approve response classes to inherit from
the create response class. This pushes the ID and status properties
to the base class and makes them all inherit from the base class. The
isSuccessful function is now abstract on the base class.

Additionally, a new function requiresApproval() is added to the
CreatePaymentResponse object. This is used to determine whether the
merchant needs to take another action on the created payment in order
to push it through.

Finally, the cancel payment status normalization is tweaked to return
FinalStatus::CANCELLED when it recieves the expected response from the
API, and CancelPaymentResponse's isSuccessful() implementation checks
for that FinalStatus constant.

Change-Id: I9530f9b62a942899f5b8f2a18ebacc4260118464