Adyen iDEAL: record payment on AUTH message
It turns out that for iDEAL payments, we never get a CAPTURE message,
so we need to send the donation to Civi on the AUTH message instead of
discarding it.

For cards, we still wait for the CAPTURE message.

Note that a few of these classes could be named more clearly.

Bug: T251296
Change-Id: Iced879a63b4fafe6cfc491e74da5c77f349cfbe2