Stop escaping Unicode for the Ingenico API
This is a part of the solution for the over-long Hungarian descriptor
and might be the entire solution for the emoji-laden names. Without
this option, we seem to be escaping any non-ascii chars with a form
that's a lot longer than their unicode equivalent. Since the Ingenico
API seems to be able to deal with UTF-8 we can just send it unescaped.

Bug: T261058
Bug: T241757
Change-Id: Id7c5ec996a03f382ab26bcfa7cfe65e4942590f1