Map AVS and CVV scores for Adyen Checkout
This is the minimum needed to pass back some normalized scores
(broken out by type, for sending to the Antifraud queue) back
up to DonationInterface. We could potentially also normalize
the codes returned from Adyen to an enum of CVV / AVS results,
then have the enum values mapped to scores in the configuration
so fraud maps could be shared across providers.

Could use the same RiskScorer for the existing IPN-responsive
ProcessCaptureRequestJob, but we can actually delete that job
when we switch to the new integration.

Bug: T282641
Change-Id: I279ce02ce3c3041a200ecb5ed577cd1c3fdabdba