Normalize refusal reasons to ErrorCodes for REST
Use same basic logic as WSDL's mapTxnIdAndErrors, just without the
'checkForRetry' bit which feels a bit strange. Perhaps the mapping
of refusal reason to error code should be handled by a separate
class, like the StatusNormalizers.

Adds an extra check on whether the raw response is null. In many
code paths, this will have already triggered an error by the time
we get to the mapRestIdAndErrors call, but for approvePayment it
helps. Updates the bad approvePayment response test to match REST
API failure modes, and fixes a copy-paste error in the same test
file.

Bug: T283117
Change-Id: I8690ded510653779d9e8fef76a32976adc7194a0