Get Apple Pay session from Apple
The request needs to authenticate itself with a certificate, so
we add cert path + password parameters to the Core/Http classes.

The Safari Apple Pay widget provides a validationUrl in the
onValidateMerchant event, so we are accepting that in the
ApplePayPaymentProvider->createPaymentSession params. The rest of
the things we post to Apple's session endpoint feel more like
config values, so I've added them as constructor parameters to
the ApplePayPaymentProvider class (as they don't apply to any
other Adyen payment methods).

For our testing setup, where we use the same merchant cert across
six domain names, we can also pass the domain name in as a param.
For other use cases it might be nicer to have that in config, so
this code can use it from either source.

For testing, please pull down the latest config-private patch and
re-run setup.sh to recreate your payments settings.

Bug: T290108
Change-Id: Icf4f3aa623bf8e51b30d4187d778e600804d0d58