Adyen iDEAL: handle 'Cancelled' payment status
I think this status is only relevant for iDEAL. We expect a different
subset of statuses on creation and on hosted payment details lookup,
so I created the new RedirectedPaymentStatus mapper.

We also make the getHostedPaymentDetails call consider different
statuses as successful depending on the concrete subclass. For the
DirectDebitPaymentProvider, Pending is considered a success.

Bug: T296266
Change-Id: I8c726b9dcc0a860ff25d09f46593ec0344f09f8b