Option to return breakdown rows for payments_fraud
Uses 'score_breakdown' as a key for a nested array for the breakdown
rows, to match the payments-antifraud queue message format.

TODO: tests. Those will need a bunch of other changes (creating
another table, etc) so I'm breaking this out for easier review.

FIXME: user_ip needs to be translated to/from binary

Bug: T291223
Change-Id: I78f9d47060960550c6ba6e349b66598525e8ef5b