PaymentsFraud: encode/decode user_ip
It needs to be converted to an integer for storage and back to a
string upon retrieval. We haven't actually been using this class
for storage in production, and the one place we use it for retrieval
we haven't been looking at the IP address, so this bug hasn't bitten
us yet.

Change-Id: I167b6610eacbff368acf2e1821bf60547f946302