Option for queue push to fall back to db
Rather than deciding at the SmashPig level where it's appropriate
to fall back to the database, this patch just adds the option to
QueueWrapper::push and lets callers opt in to the behavior. That
way we can start by using it just in the recurring charge job
and the audit parsers.

Bug: T314805
Change-Id: Ib3d87ef0b74d346abee058feb50aa776992ba84e