Support different recurring models for Adyen
Allows a new parameter 'recurring_model' to be sent in the
create payment calls for card, Google Pay, and Apple Pay. Defaults
to the Subscription model which has till now been the only
one available.

Creates a payment-provider-agnostic enum-ish class whose constants
should be used to specify the recurring_model. The string values
happen to be the same string values as Adyen expects, but this will
let us use different flags for different processors if needed.

We will need to make a matching change in DonationInterface to send
this parameter for speculatively-tokenized one-time donations (i.e.
for monthly convert).

Bug: T313854
Change-Id: If7aae9eecfd4c0b1fb806229004eebfe975284b9