Change API call to look up tokens
Adyen support explained that the /paymentMethods endpoint doesn't
return tokens for iDEAL payments, but that the /listRecurringDetails
endpoint should.

https://docs.adyen.com/api-explorer/#/Recurring/v67/post/listRecurringDetails

Note that it is on a different hostname from the /paymentMethods and
the rest of the checkout endpoints. For test, the Recurring endpoints
are on the same domain as the old WSDL endpoints, but in production
we may need to use a different value.

We recently duplicated the functionality of the getSavedPaymentDetails
call so this patch @deprecates the newer GetRecurringToken call and
delegates its function to the pre-existing getSavedPaymentDetails,
which returns a proper normalized response object with more details.

To test you can use
php PaymentProviders/Adyen/Maintenance/TestGetPaymentToken.php --pcid=xx.x

Some interesting pcids for our accounts are: 33.1 (Apple Pay),
8877665609.1 (iDEAL), and 88776655446001.1 (Visa).

Bug: T315031
Change-Id: Ic79e8eba960d3d9e5d902851dc2e736d8d0b860c