Rename IPN validator class, special-case validator
We will want to make a lot of other PayPal HTTP requests that don't
fit the EnumValidator. So we take it out of the general PayPal
config and just override it when making the single request we want
to use it for. Also renames the class which makes the API validation
requests, as it won't do any of the other interaction with the
PayPal API.

Bug: T318884
Change-Id: Ie2f9f342be27992a5d5b9a86f3cd03b85beed764