Rename 'token' params in PayPal API calls
So we can use the same parameters as we use when calling them on
Ingenico. Needed for the PayPal pending transaction resolver,
and will make it easier to use these calls with DonationInterface
where the paypal adapter normalizes the paypal-side TOKEN
parameter to gateway_session_id.

Also make the description parameter optional for the PayPal
approvePayment call, and remove an invalid PAYMENTREASON param.

Bug: T324642
Change-Id: I5cd4d00156660196a90f7f0a932d23ce06de8e74