Implement backend for DLocal recurring card payments in SmashPig
To test:
Place a donation in with this link:
https://localhost:9001/index.php?title=Special:DlocalGateway&appeal=JimmyQuote&payment_method=cc&recurring=1&uselang=en&language=en&currency=USD&amount=10&country=BR&&first_name=Lorem&last_name=Ipsum&email=test@example.com&fiscal_number=50414161173&debug=true

To find the recurring_payment_token, you can use a breakpoint on line \SmashPig\PaymentProviders\dlocal\DlocalCreatePaymentResponseFactory:23 to read the properties in the rawResponse object. The card_id is in the nested card array.

To test the createRecurringPayment for a recurring charge:

Paste the card_id gotten in the previous step in src/smashpig/PaymentProviders/dlocal/Maintenance/TestRecurringChargePayment.php:21

Then run this in the smashpig environment: php /PaymentProviders/dlocal/Maintenance/TestRecurringChargePayment.php

Bug: T324296
Depends-On: Ic0ce6e47c350e7c3588e1d754850555de480cfbf
Change-Id: I047eec180bf2579a7f16ac10d053fb3d8077a433