PayPal: Tag rather than discard GiveLively IPNs
Oops, looks like they DON'T actually import PayPal donations in the
standalone GiveLively import. Instead, we need to tag them with
the correct event code and a no_thank_you flag. Only do it for
payment messages. Note that the new place where we do the check is
guaranteed to already have a txn_type field.

Bug: T295726
Change-Id: I2299e002aff98b991a37985299d681e526a9b638