Add condition to handle braintree error that comes in a different style.
This was experienced while testing patchset 16 of this patch https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/SmashPig/+/963853 where braintree returned an error but it wasn't reflected on the client because we were not setting the error property.

This patch catches the syntax of the error and adds it to the error property of the response object.

Change-Id: I85873ad0bc46e212fd0eaadcbc2c0e1047097079