Use trait for redirect props/getters/setters
Also pull redirect members up from Create to PaymentDetailResponse

It turns out there are some cases where the PayPal getDetails call
can respond that we need to redirect the donor back to PayPal. So
we move these last unique properties up to the parent class and
leave the CreatePaymentResponse a subclass that differs in name
only.

Bug: T317245
Change-Id: If9c21b6c01fddbda8a9151b847c29fa656187fb1