Send message to cancel recurring on ended autorescue
Needs a small edit to the recurring queue consumer to make the
is_autorescue flag work in place of is_successful_autorescue.

Also tightens up the isEndedAutorescue check to make sure we
only send the message when there is a rescue reference, and
swap out the ended_auto_rescue_auth test data for an IPN that
actually DOES represent the last failed attempt.

Bug: T367451
Change-Id: I2c00f76f39308d569edec8e442ec2b17ed596f35