Remove GravyPaymentResponseFactory::fromRawResponse()
We now send in the normalized response. This patch
renames the method and adds a test to confirm
the factory behaviour.

Note: we're also detaching from IPaymentResponseFactory
here as we no longer need to build from the raw response
due to the mapping layer transforming the raw result into
a normalized version.

Bug: T364506
Change-Id: I69931351374923801a68c05d8ed0a21a5dcff401