Fix CZ online banking via Adyen
I mistakenly changed the parameter passed to rest from issuer to
issuerId in the ideal 2.0 commit referenced below. This fixes it
and adds a test to cover the mapping.

Note that I still get a response with status 'Error' from the
sandbox, but it's not the validation error that you get without this
code change.

See: If8ba39579c163e592a9460219cc88cb51edc762c
Bug: T374564
Change-Id: Icf97d2fe451fafbb10a0e58a0f593285eaaa6953