Update symfony/http-foundation
It's way out of date, and is causing security alerts.

The changelog is at
https://github.com/symfony/http-foundation/blob/7.1/CHANGELOG.md

The only bits that I think affect our code are the functions on the
Request object to set the trusted proxies (that is, the IP addresses
we trust to add a X-Forwarded-For header with the 'real' remote IP).

I made those updates in RequestHandler here, and added a patch to
config-private to add those settings to provider-defaults.

Change-Id: Iec822595a991d505f5cd7f8e296d8f2d0a8ee1e2