Refactor Gravy request mapper code.
Removed getPaymentRequest() to use getRequestMapper()
consistently across all createPayment() calls.
getRequestMapper() was added to BankPaymentProvider, which
I think is a good addition allowing us to inject
payment-method-specific mappers into the PaymentProivder
service classes. I pulled that method up to the base class
and now all Provider classes can use or extend as needed.

I also updated the visibilty of the getters to be protected
as these are implementation details that don't need to be visible
to client code.

Change-Id: I327f520a7a067a8df841347aab0bc8a875e207bb