Use redirect payment provider for Gravy Dlocal bank transfers
BankPaymentProvider has logic that was setup for direct debit payments with status "capture_pending" or "processing" to be set as successful for ingestion into civi. Given Dlocal bank transfers do not offer the 2 step process (that is Auth/Capture), this could cause donors with failing transactions to receive a "Thank you" message before the payment eventually fails.

RedirectPaymentProvider has a similar setup as BankPaymentProvider but without the extra logic for incomplete transactions.

Bug: T385012
Change-Id: Iec9299022ec8e11ee54d2f91fcfb180fd4c2102c