Fix bugs in Gravy donation queue message builder
Note: this script is now Gravy-specific. Most of the bugs
were due to trying to build a message from any processors
log data, but there are too many differences between gravy
and the other processor log data to make this feasible in
this iteration. I've renamed the file to reflect this.

Bug: T381012
Change-Id: I5078724eac6f9372d96e105ed3ac18dafde5bb49