Update PendingDatabase for is_resolved
Note that this makes the fetchMessageByGatewayOrderId
function implicitly filter by is_resolved=0 to preserve
existing expectations (see PS5 for an attempt to change
that). This feels like the safest way to move forward,
although we will have to add another function for the
audit code to use to find potentially-resolved rows.

Bug: T400304
Change-Id: Ib3e92cd04596ed2d0a7386e376608385c3e18850