Add alert suppression to ErrorTracker to prevent duplicate alerts
This patch adds a new alertSuppressionPeriod configuration parameter,
with a default of 120 seconds.

When an error threshold is breached, warranting an alert, the code now
checks Redis for a recently sent alert key before triggering a
new one. If one is found, the alert is skipped. Think of it
as an expiring lockfile.

This will ensure that the first occurrence of a threshold breach still
triggers an alert, while subsequent errors within the suppression window
will not generate failmal spam, like we got last night.

Bug: T404017
Change-Id: I8e9a727afe12e4ed900570876be2e6840842d702