Don't throw exception when no tax id needed
When a donor has a fiscal number stored but is making a payment in
a country that doesn't need us to send it, just omit the tax_id
parameter rather than throw an exception.

Note that we are currently allowing fiscal numbers that don't match
any of the supplied formats, just mapping them to the first one in
the list. We could change getGravyTaxIdTypeForSuppliedCountryIdentifier
to return null instead of $gravyCodes[0] for ones that don't match,
but I'm loathe to potentially increase failmail at this time of year!

Change-Id: Ifd7d020fbca5def3b6731f949c56d05012fd3715