Retry RecordCaptureJob if no pending data found
Also move throw call up to unindent following code. Simplifies the
if condition to just check the message exists - we have just fetched
the message by order ID, so there's no real chance it doesn't have
one set.

Also update 'delete' call and comment to reflect that we are now
just marking things as resolved to leave a trail for the audit.

Bug: T405484
Change-Id: I973fb387b85c6d72b86ea89e75196e717529fb11